backend/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.3
colorama==0.4.4
cryptography==36.0.1
DateTime==4.4
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.3
Django==4.0.5
djangorestframework==3.13.1
docopt==0.6.2
entrypoints==0.3
executing==0.8.2
filelock==3.3.1
Flask==2.1.2
Flask-Login==0.6.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
importlib-metadata==4.11.4
ipykernel==6.7.0
ipython==8.0.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.1.1
jupyter-core==4.9.1
makerchip-app==1.1.5
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.21.3
outcome==1.1.0
pandas==1.3.4
parso==0.8.3
pathspec==0.9.0
pickle-mixin==1.0.2
pickleshare==0.7.5
pipreqs==0.4.11
platformdirs==2.4.0
plotly==5.3.1
prompt-toolkit==3.0.24
pure-eval==0.2.1
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.4.0
pyOpenSSL==21.0.0
PyPDF2==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2022.1
pywin32==303
pyzmq==22.3.0
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
sseclient-py==1.7.2
stack-data==0.1.4
tenacity==8.0.1
toml==0.10.2
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
tzdata==2022.1
urllib3==1.26.7
uuid==1.30
virtualenv==20.9.0
wcwidth==0.2.5
Werkzeug==2.1.2
wsproto==1.0.0
yarg==0.1.9
zipp==3.8.0
zope.interface==5.4.0
